refac: split text slide from text paralax devider
